add extension manager tests
