+++ tests/extension/test_entrypoint.py
@@ -12,38 +12,5 @@ def test_server_extension_list(environ, script_runner):
         'server',
         'extension',
         'list',
-        env=os.environ
     )
     assert ret.success
-
-
-def test_server_extension_enable(environ, script_runner):
-    # 'mock' is not a valid extension The entry point should complete
-    # but print to sterr.
-    extension_name = "mockextension"
-    ret = script_runner.run(
-        "jupyter",
-        "server",
-        "extension",
-        "enable",
-        extension_name,
-        env=os.environ
-    )
-    assert ret.success
-    assert 'Enabling: {}'.format(extension_name) in ret.stderr
-
-
-def test_server_extension_disable(environ, script_runner):
-    # 'mock' is not a valid extension The entry point should complete
-    # but print to sterr.
-    extension_name = 'mockextension'
-    ret = script_runner.run(
-        'jupyter',
-        'server',
-        'extension',
-        'disable',
-        extension_name,
-        env=os.environ
-    )
-    assert ret.success
-    assert 'Disabling: {}'.format(extension_name) in ret.stderr
